requirements.txt
opencv-python
mediapipe==0.10.9
numpy
matplotlib
pyttsx3
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary

# Auth
PyJWT
google-auth
requests

# API
python-multipart